fix: ProjectCard height, icons and interfaces folder
refs: #8
