+++ data/projectIcons.tsx
@@ -1,17 +1,4 @@
-export interface ProjectIcons {
-  [iconName: string]: string;
-  'Adobe XD': string;
-  CSS: string;
-  Firebase: string;
-  HTML: string;
-  JavaScript: string;
-  'Material UI': string;
-  'Next.js': string;
-  'React.js': string;
-  Redux: string;
-  Sass: string;
-  Stripe: string;
-}
+import { ProjectIcons } from '../interfaces/ProjectIcons';
 
 export const projectIcons: ProjectIcons = {
   'Adobe XD': '/icons/adobexd.svg',
