+++ interfaces/AutoSliderProps.ts
@@ -1,0 +1,6 @@
+export interface AutoSliderProps {
+  perPage: number;
+  speed: number;
+  firstIndex: number;
+  lastIndex?: number;
+}
