+++ interfaces/ImageDialogProps.ts
@@ -1,0 +1,5 @@
+export interface ImageDialogProps {
+  imageUrl: string;
+  open: boolean;
+  handleClose: () => void;
+}
