+++ interfaces/ProjectIcons.ts
@@ -1,0 +1,14 @@
+export interface ProjectIcons {
+  [iconName: string]: string;
+  'Adobe XD': string;
+  CSS: string;
+  Firebase: string;
+  HTML: string;
+  JavaScript: string;
+  'Material UI': string;
+  'Next.js': string;
+  'React.js': string;
+  Redux: string;
+  Sass: string;
+  Stripe: string;
+}
