+++ interfaces/ProjectProps.ts
@@ -1,0 +1,9 @@
+export interface ProjectProps {
+  project: {
+    title: string;
+    tool: string[];
+    about: string;
+    imageUrl: string;
+    url: string;
+  };
+}
